gstplayer: move gstplayer into separate file
Finishing unifying playback element used by
audioplayers and video player plugins.
